make it a gem
