+++ Gemfile
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+source "https://rubygems.org"
+
+# Specify your gem's dependencies in convert_apple_health_export.gemspec
+gemspec
+
+gem "rake", "~> 13.0"
+gem "rspec", "~> 3.0"
+gem "solargraph"
+gem "solargraph-standardrb"
+gem "standard", "~> 1.3.0"
