+++ exe/convert_apple_health_export
@@ -1,0 +1,6 @@
+#!/usr/bin/env ruby
+
+$LOAD_PATH.unshift File.expand_path("../lib/", File.dirname(__FILE__))
+require "convert_apple_health_export"
+
+ConvertXML.call("export.xml", "export.csv")
