+++ lib/convert_apple_health_export/convert_xml.rb
@@ -1,0 +1,12 @@
+module ConvertXML
+  extend self
+
+  def call(input_path, export_path)
+    diastolic_records, systolic_records = ParseXML.call(input_path)
+    records = JoinRecords.call(diastolic_records, systolic_records)
+
+    puts "Found #{records.count} records, creating CSV"
+
+    CreateCSV.call(records, export_path)
+  end
+end
