+++ lib/convert_apple_health_export/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module ConvertAppleHealthExport
+  VERSION = "0.1.0"
+end
