+++ spec/convert_apple_health_export_spec.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+RSpec.describe ConvertAppleHealthExport do
+  it "has a version number" do
+    expect(ConvertAppleHealthExport::VERSION).not_to be nil
+  end
+
+  it "does something useful" do
+    expect(false).to eq(true)
+  end
+end
